backend/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-silk==5.0.4
djangorestframework==3.14.0
drf-yasg==1.21.7
gprof2dot==2022.7.29
gunicorn==21.2.0
inflection==0.5.1
numpy==1.26.2
packaging==23.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pymemcache==4.0.0
python-memcached==1.59
pytz==2023.3.post1
PyYAML==6.0.1
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
whitenoise==6.6.0
